+++ cdisc_rules_engine/services/cache/__init__.py
@@ -1,10 +1,8 @@
 from .cache_service_factory import CacheServiceFactory
-from .cache_service_interface import CacheServiceInterface
 from .in_memory_cache_service import InMemoryCacheService
 from .redis_cache_service import RedisCacheService
 
 __all__ = [
-    "CacheServiceInterface",
     "CacheServiceFactory",
     "InMemoryCacheService",
     "RedisCacheService",
