+++ cdisc_rules_engine/services/data_services/base_data_service.py
@@ -7,9 +7,9 @@ import pandas as pd
 
 from cdisc_rules_engine.config import ConfigService
 from cdisc_rules_engine.constants.domains import AP_DOMAIN_LENGTH
+from cdisc_rules_engine.interfaces import CacheServiceInterface
 from cdisc_rules_engine.models.dataset_types import DatasetTypes
 from cdisc_rules_engine.services import logger
-from cdisc_rules_engine.services.cache import CacheServiceInterface
 from cdisc_rules_engine.utilities.utils import (
     get_dataset_cache_key_from_path,
     get_directory_path,
