+++ cdisc_rules_engine/services/data_services/dummy_data_service.py
@@ -5,7 +5,7 @@ import pandas as pd
 from cdisc_rules_engine.config import ConfigService
 from cdisc_rules_engine.dummy_models.dummy_dataset import DummyDataset
 from cdisc_rules_engine.exceptions.custom_exceptions import DatasetNotFoundError
-from cdisc_rules_engine.services.cache import CacheServiceInterface
+from cdisc_rules_engine.interfaces import CacheServiceInterface
 from cdisc_rules_engine.services.data_services import BaseDataService
